Stub out the basic purge hooks in fastly/hooks
The hooks are currently implemented as print statements of what
they would do, rather than actually being done.

In limited testing of this realized that R: and RT: key handling
was incorrect, for reason of policy changes on recipes. Therefore
RT: is removed and R: is used in more places.